refactor: rename status file
